Address issue /issues/16 by making exercises more explicit.
In particular, prescribe a clustering method/key function that does not eliminate leading and trailing blanks.  Also, divide the last section into three smaller steps: Split, Undo/Redo, and Trim Leading and Trailing Whitespace. The intention is to reduce cognitive load on the learner, but a review exercise is slipped into the last exercise to make sure they're not getting it too easy.